+++ src/main/java/bigbrother/slimdealz/config/EnvConfig.java
@@ -1,9 +1,9 @@
-//package bigbrother.slimdealz.config;
-//
-//import org.springframework.context.annotation.Configuration;
-//import org.springframework.context.annotation.PropertySource;
-//
-//@Configuration
-//@PropertySource("classpath:properties/env.properties")
-//public class EnvConfig {
-//}
+package bigbrother.slimdealz.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+
+@Configuration
+@PropertySource("classpath:properties/env.properties")
+public class EnvConfig {
+}
