--- src/main/java/bigbrother/slimdealz/dto/JwtRequestDto.java
@@ -1,4 +1,0 @@
-package bigbrother.slimdealz.dto;
-
-public class JwtRequestDto {
-}
