--- src/main/java/bigbrother/slimdealz/dto/NotificationDto.java
@@ -1,4 +1,0 @@
-package bigbrother.slimdealz.dto;
-
-public class NotificationDto {
-}
