+++ src/main/java/bigbrother/slimdealz/dto/product/PriceHistoryDto.java
@@ -1,4 +1,4 @@
-package bigbrother.slimdealz.dto;
+package bigbrother.slimdealz.dto.product;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,10 +9,9 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class PriceDto {
+public class PriceHistoryDto {
     private Long id;
-    private int setPrice;
-    private String promotion;
-    private Long productId;
-    private VendorDto vendor;
+    private int previousPrice;
+    private String updatedAt;
+    private String endAt;
 }
