+++ src/main/java/bigbrother/slimdealz/dto/product/ProductDto.java
@@ -1,4 +1,4 @@
-package bigbrother.slimdealz.dto;
+package bigbrother.slimdealz.dto.product;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
