+++ src/main/java/bigbrother/slimdealz/dto/user/NotificationDto.java
@@ -1,0 +1,4 @@
+package bigbrother.slimdealz.dto.user;
+
+public class NotificationDto {
+}
